--- src/hooks/use-app-dispatch.ts
@@ -1,6 +1,0 @@
-import { useDispatch } from 'react-redux';
-import { AppDispatch } from '../redux/store';
-
-const useAppDispatch = () => useDispatch<AppDispatch>();
-
-export default useAppDispatch;
