--- src/hooks/use-app-selector.ts
@@ -1,6 +1,0 @@
-import { useSelector, TypedUseSelectorHook } from 'react-redux';
-import { RootState } from '../store/theme-slice';
-
-const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
-
-export default useAppSelector;
