+++ src/main.tsx
@@ -2,7 +2,7 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 import { Provider } from 'react-redux';
-import store from './redux/store';
+import store from './store/store';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <Provider store={store}>
